requirements.txt
colorama>=0.4.6
dnspython>=2.3.0
requests>=2.28.2
tqdm>=4.64.1
urllib3>=1.26.15
jinja2>=3.1.2 